make modified file detection more lazy
