+++ airbyte-ci/connectors/pipelines/pipelines/airbyte_ci/connectors/commands.py
@@ -11,7 +11,10 @@ from connector_ops.utils import ConnectorLanguage, SupportLevelEnum, get_all_con
 from pipelines import main_logger
 from pipelines.cli.click_decorators import click_append_to_context_object, click_ignore_unused_kwargs, click_merge_args_into_context_obj
 from pipelines.cli.lazy_group import LazyGroup
+from pipelines.consts import CIContext
 from pipelines.helpers.connectors.modifed import ConnectorWithModifiedFiles, get_connector_modified_files, get_modified_connectors
+from pipelines.helpers.git import get_modified_files_in_commit, get_modified_files_in_branch
+from pipelines.helpers.utils import transform_strs_to_paths
 
 ALL_CONNECTORS = get_all_connectors_in_repo()
 
@@ -47,10 +50,6 @@ def get_selected_connectors_with_modified_files(
     Returns:
         List[ConnectorWithModifiedFiles]: The connectors that match the selected criteria.
     """
-
-    if metadata_changes_only and not modified:
-        main_logger.info("--metadata-changes-only overrides --modified")
-        modified = True
 
     selected_modified_connectors = (
         get_modified_connectors(modified_files, ALL_CONNECTORS, enable_dependency_scanning) if modified else set()
@@ -236,6 +235,19 @@ async def connectors(
     """Group all the connectors-ci command."""
     validate_environment(ctx.obj["is_local"])
 
+    if ctx.obj["metadata_changes_only"] and not ctx.obj["modified"]:
+        main_logger.info("--metadata-changes-only overrides --modified")
+        ctx.obj["modified"] = True
+
+    if ctx.obj["modified"]:
+        ctx.obj["modified_files"] = transform_strs_to_paths(await get_modified_files(
+            ctx.obj["git_branch"],
+            ctx.obj["git_revision"],
+            ctx.obj["diffed_branch"],
+            ctx.obj["is_local"],
+            ctx.obj["ci_context"],
+        ))
+
     ctx.obj["selected_connectors_with_modified_files"] = get_selected_connectors_with_modified_files(
         ctx.obj["names"],
         ctx.obj["support_levels"],
@@ -243,7 +255,22 @@ async def connectors(
         ctx.obj["modified"],
         ctx.obj["metadata_changes_only"],
         ctx.obj["metadata_query"],
-        ctx.obj["modified_files"],
+        set(ctx.obj["modified_files"]),
         ctx.obj["enable_dependency_scanning"],
     )
     log_selected_connectors(ctx.obj["selected_connectors_with_modified_files"])
+
+
+async def get_modified_files(git_branch: str, git_revision: str, diffed_branch: str, is_local: bool, ci_context: CIContext) -> Set[str]:
+    """Get the list of modified files in the current git branch.
+    If the current branch is master, it will return the list of modified files in the head commit.
+    The head commit on master should be the merge commit of the latest merged pull request as we squash commits on merge.
+    Pipelines like "publish on merge" are triggered on each new commit on master.
+
+    If the CI context is a pull request, it will return the list of modified files in the pull request, without using git diff.
+    If the current branch is not master, it will return the list of modified files in the current branch.
+    This latest case is the one we encounter when running the pipeline locally, on a local branch, or manually on GHA with a workflow dispatch event.
+    """
+    if ci_context is CIContext.MASTER or (ci_context is CIContext.MANUAL and git_branch == "master"):
+        return await get_modified_files_in_commit(git_branch, git_revision, is_local)
+    return await get_modified_files_in_branch(git_branch, git_revision, diffed_branch, is_local)
